# Speed up simulation