--- a/tests/test_baseunit.py
+++ b/tests/test_baseunit.py
@@ -107,7 +107,7 @@ def test_calculate_bids(base_unit, mock_market_config):
     # mock calculate_marginal_cost
     base_unit.calculate_marginal_cost = lambda *x: 10
     base_unit.set_dispatch_plan(mock_market_config, orderbook)
-    base_unit.calculate_generation_cost(index[0], index[1], "energy")
+    base_unit.calculate_generation_cost(index[0], index[-1], "energy")
 
     # we apply the dispatch plan of 10 MW
     assert base_unit.outputs["energy"][start] == 10
@@ -117,7 +117,7 @@ def test_calculate_bids(base_unit, mock_market_config):
 
     # we somehow sold an additional 10 MW
     base_unit.set_dispatch_plan(mock_market_config, orderbook)
-    base_unit.calculate_generation_cost(index[0], index[1], "energy")
+    base_unit.calculate_generation_cost(index[0], index[-1], "energy")
 
     # the final output should be 10+10
     assert base_unit.outputs["energy"][start] == 20
@@ -153,7 +153,7 @@ def test_calculate_multi_bids(base_unit, mock_market_config):
     # mock calculate_marginal_cost
     base_unit.calculate_marginal_cost = lambda *x: 10
     base_unit.set_dispatch_plan(mock_market_config, orderbook)
-    base_unit.calculate_generation_cost(index[0], index[1], "energy")
+    base_unit.calculate_generation_cost(index[0], index[-1], "energy")
 
     assert base_unit.outputs["energy"][index[0]] == 10
     assert base_unit.outputs["energy_marginal_costs"][index[0]] == 100
@@ -163,7 +163,7 @@ def test_calculate_multi_bids(base_unit, mock_market_config):
     assert base_unit.outputs["energy_cashflow"][index[1]] == 110
 
     base_unit.set_dispatch_plan(mock_market_config, orderbook)
-    base_unit.calculate_generation_cost(index[0], index[1], "energy")
+    base_unit.calculate_generation_cost(index[0], index[-1], "energy")
 
     # should be correctly applied for the sum, even if different hours are applied
     assert base_unit.outputs["energy"][index[0]] == 20