-refactor the function itself to use vectorization
-execute the funtion only once at the end of the simulation
-paralellize the call for all agents to improve performance

suggestion by @maurerle


The speed up is as follows (tested on small example):
current version - 24s
vectorized calculate_generation_cost function - 16s
execute calculation only at the end of the simulation - 8s

Using the small_learning_1 example:
current version - 36s
final tweaked version - 10s